fix(退租金)

+++ src/main/java/com/xiliulou/electricity/entity/BatteryMemberCard.java
@@ -45,7 +45,7 @@ public class BatteryMemberCard {
     /**
      * 租期
      */
-    private Integer validDays;
+    private Long validDays;
     /**
      * 租期单位 0：分钟，1：天
      */
